Update Cray compiler workarounds
There's multiple facets to these changes to fix CCE handling allocatable
module arrays:

1) CCE only has a problem with allocatable module arrays, not scalars or
statically sized arrays. Use declare create for everything that isn't allocatable.
2) Allocatable array handling is broken in CCE <= 16.0.0, but an effective
(but ugly) workaround is to add a pointer in front of the allocatable
and leverage the pointer attachment logic to fixup the link.
The CRAY_DECLARE_GLOBAL macro declares the shadow allocatable and the pointer.
The ALLOCATE_GLOBAL macro allocates the shadow, attaches the pointer,
and creates/attaches each on the device.
The DEALLOCATE_GLOBAL macro detaches and releases the device entries,
nullifies the pointer, and deallocates the shadow.
The ALLOCATE and DEALLOCATE macros can be used for local variables or
derived type components.

This commit still isn't functional on AMD GPUs due to some register allocation
issues in a specific function. Those will be addressed in a future commit.