TBR: give global private scalars a unique name
It's possible that CCE is not mangling the private global names on the device
in a way that makes them unique, which could be causing problems.

Most of these don't need to declared to the device, though.
They're loop bounds which shouldn't need to be mapped.

Need to check if these WARs are actually neeeded.