[cray] Necessary changes to remove the CRAY_ACC_WAR
The main thrust of this commit factors out certian procedure pointers that are causing a bug in the Cray compiler that ignores all `acc declare` statements. This is annoying workaround but not catastrophic given that procedure pointers could be refactored into dispatch subroutines easily and runtime polymorpism was minimal.

There are a few other assorted tweaks in the non-CRAY_ACC_WAR logic.

WARNING: more complicated boundary conditions are broken in this commit! The next commit will fix them.