Changes to through the code to make fluid variables uniform
The following PR comprises changes to the code in order to call fluid variables uniformly through the code. The main changes for this PR is to change all the fluid_pp(i)% to their respective, which are defined on src/common/m_variables_conversion.fpp. For instance, substituting fluid_pp(i)%pi_inf by pi_infs(i), and so on.

In certain cases, variable indices are also replaced to be uniform

The importance of this PR is that, if these variables change through the code for whatever reason, they will be followed through the rest of the code